+++ src/main/java/com/gxf/his/service/DepartmentService.java
@@ -13,6 +13,6 @@ public interface DepartmentService {
 
     void addDepartment(Department department);
 
-    void deleteDepartment(Integer departmentId);
+    void deleteDepartment(Long departmentId);
 
 }
